chore(#50): uncoupling CalculateAppprovalsRate from database
