+++ src/@types/appTypes/ApprovalsAndDisapprovals.d.ts
@@ -1,0 +1,5 @@
+export default interface ApprovalsAndDisapprovals
+{
+  approvals: number;
+  disapprovals: number;
+};
